refactor(test): split testPersonas into two separate files
