+++ package.json
@@ -11,7 +11,7 @@
 		"compile:personas": "node scripts/personasToJSON.js",
 		"compile:rules": "node scripts/rulesToJSON.mjs",
 		"compile-watch": "nodemon -e yaml --watch data/ scripts/rulesToJSON.mjs & nodemon -e yaml --watch personas/ scripts/personasToJSON.js & nodemon -e md --watch contenu-ecrit/ scripts/documentationToJSON.js ",
-		"test": "yarn compile:personas && node tests/testPersonas.js",
+		"test": "yarn compile:personas && yarn compile:rules && node tests/testPersonas.js",
 		"test:md": "node tests/testPersonas.js --markdown",
 		"translate": "node scripts/i18n/translate-rules.js && node scripts/i18n/translate-personas.js",
 		"check:personas": "node scripts/i18n/check-personas.js ",
